Better delimiting between commands
We have a UICommand type that the client will send to the UI as a side-effect of admin *events*.

It is better to think of admin's messages as events rather than commands as it seems more natural.

In this way the event can be send to all player clients, i.e. broadcasted, and the appropriate player client
can respond.

All events are handled by a path of form /event/<event_name> by the player client.

Implementing the served_cards event in admin and player client. This can be seen as an event
that all player clients respond to in the same manner - syncing the table.

CONSIDER(@rk): Does it make sense to implement a player-to-player "check table" command?
Will help with testing.